Player chooses to switch lanes when arriving at a junction.
Made a new class called TrackJunction, and new buttons Track1,2. can add more tracks in the future. when arriving at a junction an empty object collider is triggered to stop the car and then based on which track is chosen the car proceeds. will work on making the whole aabb state diagram